further additions as i muddle through fight.c and add all of the needed stuff

+++ src/main/java/com/wallyatkins/mud/merc/types/ApplyAffectType.java
@@ -34,7 +34,7 @@ public enum ApplyAffectType {
 		this.index = index;
 	}
 	
-	public int index() {
+	public short index() {
 		return index;
 	}
